Fixed pdb/idb bug: MT build goes completely mad if projects mutually overwrite incremental rebuild database. Restructured cluster inheritance hierarchy minimising multiple inheritance. Added some convenience cluster classes. Addes cluster tests. Fixed pointer to int warning.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@565 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ base/include/vislib/RegEx.inl
@@ -765,7 +765,7 @@ typename vislib::RegEx<T>::Expression *vislib::RegEx<T>::parseRepeatEx(
     }
     try {
         minRep = T::ParseInt(String<T>(begin, 
-            VRECTX_CUR_STR(inOutCtx) - begin));
+            static_cast<int>(VRECTX_CUR_STR(inOutCtx) - begin)));
         ASSERT(VRECTX_CUR_STR(inOutCtx) != begin);
     } catch (...) {
         throw ParseException(ParseException::INTEGER_EXPECTED, begin, __FILE__,
@@ -795,7 +795,7 @@ typename vislib::RegEx<T>::Expression *vislib::RegEx<T>::parseRepeatEx(
     }
     try {
         maxRep = T::ParseInt(String<T>(begin, 
-            VRECTX_CUR_STR(inOutCtx) - begin));
+            static_cast<int>(VRECTX_CUR_STR(inOutCtx) - begin)));
         ASSERT(VRECTX_CUR_STR(inOutCtx) != begin);
     } catch (...) {
         throw ParseException(ParseException::INTEGER_EXPECTED, begin, __FILE__,
